Win/Loss Screen update
Moved the end screen from being separate scenes to act within the main scene itself via panels. Tweaked a few scripts as needed to add references to the objects and commented out redundant lines of code, and lastly added new lines that updated the score on every instance of the tracker, even when disabled.